reorganize tests into files
